[chore] Organize query and schema files
