+++ poe/urls.py
@@ -18,5 +18,5 @@ from django.urls import path
 from graphene_django.views import GraphQLView
 
 urlpatterns = [
-    path("graphql", GraphQLView.as_view(graphiql=True)),
+    path("graphql/", GraphQLView.as_view(graphiql=True)),
 ]
